backup_v1_stable/requirements.txt
Flask
gunicorn
yfinance
pandas
numpy
requests
google-generativeai
